refactor(agentic): split agent-detection-service into 5 slices
Decomposes the 1300-line monolith into focused modules that mirror the
branch-lifecycle layout from the prior cycle:

- agent-detection-registry: canonical DetectedAgent list + agentsStore;
  internal `_registry` mutation accessor
- agent-detection-state-store: per-pane reactive stores
  (paneAgentTypeStore + paneAgentStateStore) and AgentState dispatcher
- agent-detection-patterns: pure pattern matching, DEFAULT_PATTERNS,
  surface → workspace / pty / argv lookups
- agent-detection-publish: StatusTracker + attach / detach / publish
  lifecycle that mutates registry, pane store, status registry, event
  bus, and surface-service unread flag
- agent-detection-controller: initAgentDetection / destroy / reset and
  all event-bus + OSC subscriptions; consolidates the duplicate argv →
  intendedAgent fallback into `seedPaneAgentTypeIfAbsent`

agent-detection-service.ts is now a barrel re-exporting the historical
public surface so component / extension / test imports stay stable.
